Improved gleaning of tiff time from granule name
Use time entry in dimensions dict

Ex:
```yaml
time:
  field:
    delimiter: "_"
    n: 5
  index:
    start: 20
    end: 35
  format: "%Y%m%dT%H%M%S"
```

One of either `time.field` or `time.index` and `time.format` are required.

`time.field`:
`delimiter` is a regex that splits the granule filename
`n` is the index of the field with the time data following the split

`time.index`:
`start` and `end` give the start and end indices of the time string. Used as `granule_name[start:end]`

`format` is the format of the time field, as defined in `datetime.strptime`